Refactor into smaller focused modules.
